Hadrian: install extra dynamic flavours of libHSrts
Summary:
Previously, Hadrian was building all the appropriate dynamic ways for libHSrts
but they were not picked up and installed in the package database when
we register the rts library. Since we use Cabal for registering packages and
the .cabal files of packages as sources of truth for configuring and installing,
we ended up patching Cabal to add a new field, 'extra-dynamic-library-flavours',
to specify those extra flavours to install in .cabal files:

    https://github.com/haskell/cabal/pull/5606

We now make use of this in rts.cabal.in to expose dynamic flavours behind a
Cabal flag, which Hadrian will use whenever we are building a GHC flavour that
requires dynamic libraries.

This is all part of a larger plan to build a dynamic stage 2 GHC by default,
like with make, which in turn will fix a lot of test failures. See #15837.

Test Plan:
hadrian/build.sh _build/stage1/lib/package.conf.d/rts-1.0.conf
_build/stage1/lib/x86_64-.../ should contain many libHSrts-*.so

Reviewers: snowleopard, DavidEichmann, bgamari, erikd, simonmar

Reviewed By: snowleopard, DavidEichmann

Subscribers: rwbarton, carter

GHC Trac Issues: #15837

Differential Revision: https://phabricator.haskell.org/D5385